@@ -1,6 +1,6 @@
 package com.apispring.moviesapi.entities;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
